fixed for optional java unzip install listall fix param added in nb.conf

+++ scripts/servers_manager_remove.sh
@@ -1,14 +1,25 @@
 echo "Removing Server - Manager"
+removeJava=$1
+#echo "removeJava :"$removeJava
+removeUnzip=$2
+#echo "removeUnzip :"$removeUnzip
+
 homeDir=$(pwd)
 source setenv.sh
 #sudo su
-yum -y remove java*
-yum -y remove jdk*
-echo "Java Remove -Done!"
-yum -y remove unzip
-echo "unzip Remove -Done!"
-yum -y remove wget
-echo "wget Remove -Done!"
+if [ "$removeJava" == "y" ]; then
+  echo "Removing Java"
+  yum -y remove java*
+  yum -y remove jdk*
+  echo "Java Remove -Done!"
+fi
+if [ "$removeUnzip" == "y" ]; then
+  echo "Removing Unzip"
+  yum -y remove unzip
+  echo "unzip Remove -Done!"
+fi
+#yum -y remove wget
+#echo "wget Remove -Done!"
 #rm -r install/*.zip
 source setenv.sh
 systemctl stop gs.service
